Add routing for tea education
