+++ src/components/app/App.tsx
@@ -4,6 +4,7 @@ import Nav from '../nav/Nav';
 import Home from '../home/Home';
 import Teas from '../teas/Teas';
 import Error from '../error/Error'
+import TeaEd from '../teaEd/TeaEd'
 import { Routes, Route} from 'react-router-dom';
 import { useState} from 'react';
 
@@ -18,7 +19,7 @@ function App() {
       <main>
         <Routes>
           <Route path='/' element={<Home isRedirected={isRedirected} setIsRedirected={setIsRedirected}/>}/>
-          <Route/>
+          <Route path='/tea/:category/education' element={<TeaEd />}/>
           <Route path='/tea/:category' element={<Teas setIsRedirected={setIsRedirected}/>}/>
           <Route path='*' element={<Error isRedirected={isRedirected} setIsRedirected={setIsRedirected}/>}/> 
         </Routes>
